requirements.txt
Flask
requests
numpy
pandas
scikit-learn
joblib
python-dotenv
google-generativeai
gunicorn